Issue #119: Use token replacement in XML feed
Jekyll doesn't support referencing variables in YAML front matter, which
means we can't do something like:

```
drupal_planet_summary: [Example]({{ site.url }}/example)
```

This is important because we need absolute links for our XML feed when
we reference internal pages or images.

As a work around, in the `feed-drupal.xml` template, we're replacing the
`!SITE_URL` token with the contents of `{{ site.url }}`. Hackish, but
it works.

The new syntax is:

```
drupal_planet_summary: [Example](!SITE_URL/example)
```

See
http://stackoverflow.com/questions/14487110/include-jekyll-liquid-template-data-in-a-yaml-variable?lq=1
for more information