Issue #241: Remove site.baseurl variable
The site.baseurl variable is designed for sites that have a path as a
prefix, for example, if we had a site called Jazzed at
http://savaslabs.github.io/jazzed, we would use the site.baseurl
variable to specify correct paths to images and local links.

That's not the case for the site, so we can remove site.baseurl, as it's
just causing confusion.